+++ src/javascript/8000-Global.js
@@ -1,0 +1,55 @@
+/*
+Global objects scattered throughout code:
+----
+from Node.js:
+    s.nodes : Map(<<SC Passage obj>> : Node)
+    s.getNode(psgTitle: String): Node
+    s.specialPassages : String[]
+    v.embeddedPsgs : String[]
+*/
+
+Config.passages.nobr = true; // sets all passages to `nobr`
+
+window.onerror = function(msg, url, linenumber) {
+    /*
+    Ensures that errors appear in a pop-up for greater visibility. Works
+    in Firefox, but doesn't seem to work in Safari. Be sure to do
+    testing in Firefox.
+    */
+    alert(
+        'Error message: ' + msg + '\n' +
+        'URL: ' + url + '\n' +
+        'Line Number: ' + linenumber
+    );
+    return true;
+}
+
+s.makeLink = function(startPsgTitle, text, endPsgTitle, func) {
+    /*
+    Calls the `addLink` method of the node associated with
+    `startPsgTitle`, passing it the other parameters. If no such node
+    exists, one is created.
+    */
+    var startNode = s.nodes.get(Story.get(startPsgTitle));
+    if (startNode === undefined) {
+        startNode = new s.Node(startPsgTitle);
+    }
+
+    return startNode.addLink(text, endPsgTitle, func);
+}
+
+v.parser = new s.Parser();
+
+Config.passages.onProcess = function (p) {
+    var processedText = v.parser.procAllMarkup(p.title, p.text);
+    return processedText;
+};
+
+v.page = new s.Page();
+
+s.rebuildPage = function(ev) {
+    v.page.rebuildPage(ev.passage);
+    return;
+}
+
+$(document).on(':passagedisplay', s.rebuildPage);